# docs/swagger.yaml
basePath: /
definitions:
...
  dto.BookingData:
    properties:
      admin_notes:
        type: string
      booking_date:
        type: string
      created_at:
        type: string
      duration_hours:
        type: integer
      end_time:
        type: string
      id:
        type: integer
      start_time:
        type: string
      status:
        type: string
      studio:
        $ref: '#/definitions/dto.StudioData'
      studio_id:
        type: integer
      total_price:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/dto.UserData'
      user_id:
        type: integer
    type: object
  dto.BookingListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.BookingData'
        type: array
      meta:
        $ref: '#/definitions/dto.PaginationMeta'
      success:
        type: boolean
    type: object
  dto.BookingResponse:
    properties:
      data:
        $ref: '#/definitions/dto.BookingData'
      success:
        type: boolean
    type: object
  dto.CancelBookingRequest:
    properties:
      reason:
        minLength: 10
        type: string
    required:
    - reason
    type: object
  dto.CancelBookingResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
...
  dto.CreateBookingRequest:
    properties:
      booking_date:
        description: YYYY-MM-DD
        type: string
      duration_hours:
        type: integer
      end_time:
        description: HH:MM
        type: string
      start_time:
        description: HH:MM
        type: string
      studio_id:
        type: integer
    required:
    - booking_date
    - end_time
    - start_time
    - studio_id
    type: object
  dto.CreateBookingResponse:
    properties:
      data:
        $ref: '#/definitions/dto.BookingData'
      message:
        type: string
      success:
        type: boolean
    type: object
...
  dto.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
      success:
        type: boolean
    type: object
  dto.LoginData:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/dto.UserData'
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      data:
        $ref: '#/definitions/dto.LoginData'
      message:
        type: string
      success:
        type: boolean
    type: object
...
  dto.PaginationMeta:
    properties:
      current_page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  dto.ProfileResponse:
    properties:
      data:
        $ref: '#/definitions/dto.UserData'
      success:
        type: boolean
    type: object
  dto.RegisterRequest:
    properties:
      email:
        type: string
      name:
        minLength: 2
        type: string
      password:
        minLength: 6
        type: string
      password_confirmation:
        type: string
    required:
    - email
    - name
    - password
    - password_confirmation
    type: object
  dto.RegisterResponse:
    properties:
      data:
        $ref: '#/definitions/dto.UserData'
      message:
        type: string
      success:
        type: boolean
    type: object
  dto.StudioData:
    properties:
      created_at:
        type: string
      description:
        type: string
      facilities:
        items:
          type: string
        type: array
      id:
        type: integer
      image_url:
        type: string
      is_active:
        type: boolean
      location:
        type: string
      name:
        type: string
      operating_hours:
        type: string
      price_per_hour:
        type: integer
      updated_at:
        type: string
    type: object
...
  dto.UpdateBookingStatusRequest:
    properties:
      admin_notes:
        description: Catatan pembayaran/perubahan status
        type: string
      status:
        enum:
        - pending
        - confirmed
        - completed
        - cancelled
        type: string
    required:
    - status
    type: object
  dto.UpdateBookingStatusResponse:
    properties:
      data:
        $ref: '#/definitions/dto.BookingData'
      message:
        type: string
      success:
        type: boolean
    type: object
...
  dto.UserData:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Dokumentasi API Backend Booking Studio
  title: Backend Booking Studio API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login dengan email & password, mengembalikan JWT token
      parameters:
      - description: Data login user
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Invalid request payload / kredensial salah
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Login user
      tags:
      - Auth
  /auth/profile:
    get:
      consumes:
      - application/json
      description: Mengambil data profil user berdasarkan JWT token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProfileResponse'
        "401":
          description: Unauthorized / token invalid
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ambil profil user login
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: 'Membuat akun baru (default role: customer)'
      parameters:
      - description: Data registrasi user
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RegisterResponse'
        "400":
          description: Invalid request payload / validasi gagal
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Register user baru
      tags:
      - Auth
  /bookings:
    get:
      consumes:
      - application/json
      description: Mengambil semua booking milik user login
      parameters:
      - description: Filter status
        in: query
        name: status
        type: string
      - description: Halaman
        in: query
        name: page
        type: integer
      - description: Jumlah per halaman
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BookingListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ambil riwayat booking user
      tags:
      - Bookings
    post:
      consumes:
      - application/json
      description: Customer membuat booking studio
      parameters:
      - description: Data booking
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.CreateBookingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreateBookingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Buat booking baru
      tags:
      - Bookings
  /bookings/{id}:
    get:
      consumes:
      - application/json
      description: Customer/Admin bisa lihat detail booking
      parameters:
      - description: ID Booking
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BookingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ambil detail 1 booking
      tags:
      - Bookings
  /bookings/{id}/cancel:
    post:
      consumes:
      - application/json
      description: Customer membatalkan booking yang masih aktif
      parameters:
      - description: ID Booking
        in: path
        name: id
        required: true
        type: integer
      - description: Alasan pembatalan
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.CancelBookingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CancelBookingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Batalkan booking
      tags:
      - Bookings
  /bookings/admin:
    get:
      consumes:
      - application/json
      description: Mengambil semua booking dengan filter
      parameters:
      - description: Filter status
        in: query
        name: status
        type: string
      - description: Halaman
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BookingListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ambil semua booking (Admin Only)
      tags:
      - Bookings
  /bookings/admin/{id}/status:
    put:
      consumes:
      - application/json
      description: Admin mengubah status booking
      parameters:
      - description: ID Booking
        in: path
        name: id
        required: true
        type: integer
      - description: Status baru
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateBookingStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UpdateBookingStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update status booking (Admin Only)
      tags:
      - Bookings
...
swagger: "2.0"